requirements.txt
fastapi==0.110.0

gunicorn==21.2.0
uvicorn[standard]==0.27.1

requests
msal
python-dotenv
azure-storage-blob
pandas
openpyxl
